Put the paths for application-specific stuff into a custom class.
Python's scoping is either kind of tricky, or the interpreter and
various linters/checkers disagree. I have almost exclusively coded in
Go and a little bit C# for the past decade, so I'm still getting
re-familiarized.
I used to do a lot of stuff in Python, but that was a long time
ago. Plus I never touched Python 3 back then.